+++ src/interactiveBlock.cpp
@@ -2,7 +2,7 @@
 #include "blockIdleState.h"
 #include <cstdio>
 
-InteractiveBlock::InteractiveBlock(SDL_Renderer *renderer, int x, int y, int w, int h, std::vector<SDL_Texture*> entityTextures, int id, CollitionSystem* collisionSystem)
+InteractiveBlock::InteractiveBlock(SDL_Renderer *renderer, float x, float y, int w, int h, std::vector<SDL_Texture*> entityTextures, int id, CollitionSystem* collisionSystem)
 : Block(renderer, x, y, w, h, entityTextures, id, collisionSystem) 
 {
     originalY = y;
